requirements.txt
django
google-generativeai
django-environ
PyPDF2
reportlab